+++ resources/js/addon.js
@@ -1,3 +1,0 @@
-if (typeof Statamic.$bard.tiptap.extensions !== 'undefined') {
-    require('./boot');
-}