+++ src/Marks/Code.php
@@ -2,9 +2,9 @@
 
 namespace JackSleight\StatamicBardMutator\Marks;
 
-use JackSleight\StatamicBardMutator\Support\Traits\MutatesMark;
+use JackSleight\StatamicBardMutator\Traits\Mutates;
 
-class Code extends \ProseMirrorToHtml\Marks\Code
+class Code extends \Tiptap\Marks\Code
 {
-    use MutatesMark;
+    use Mutates;
 }