+++ src/Marks/Link.php
@@ -2,9 +2,9 @@
 
 namespace JackSleight\StatamicBardMutator\Marks;
 
-use JackSleight\StatamicBardMutator\Support\Traits\MutatesMark;
+use JackSleight\StatamicBardMutator\Traits\Mutates;
 
 class Link extends \Statamic\Fieldtypes\Bard\LinkMark
 {
-    use MutatesMark;
+    use Mutates;
 }