+++ src/Marks/Superscript.php
@@ -2,9 +2,9 @@
 
 namespace JackSleight\StatamicBardMutator\Marks;
 
-use JackSleight\StatamicBardMutator\Support\Traits\MutatesMark;
+use JackSleight\StatamicBardMutator\Traits\Mutates;
 
-class Superscript extends \ProseMirrorToHtml\Marks\Superscript
+class Superscript extends \Tiptap\Marks\Superscript
 {
-    use MutatesMark;
+    use Mutates;
 }