+++ src/Nodes/Blockquote.php
@@ -2,9 +2,9 @@
 
 namespace JackSleight\StatamicBardMutator\Nodes;
 
-use JackSleight\StatamicBardMutator\Support\Traits\MutatesNode;
+use JackSleight\StatamicBardMutator\Traits\Mutates;
 
-class Blockquote extends \ProseMirrorToHtml\Nodes\Blockquote
+class Blockquote extends \Tiptap\Nodes\Blockquote
 {
-    use MutatesNode;
+    use Mutates;
 }